[Type System]: Mutability is now a feature of the type system, not the variable system

+++ oc/compiler/parser/parser.h
@@ -72,7 +72,7 @@ struct front_end_results_package_t{
 /**
  * For printing formatted parser errors
  */
-void print_parse_message(parse_message_type_t message_type, char* info, u_int16_t line_num);
+void print_parse_message(parse_message_type_t message_type, char* info, u_int32_t line_num);
 
 /**
  * Parse the entirety of the file. Returns 0 if successful
